db 파일생성 및 테이블 생성 구현
