+++ .gitignore
@@ -29,3 +29,4 @@ dist-ssr
 
 # sales_note
 package-lock.json
+*.sqlite
