+++ vite.config.js
@@ -8,4 +8,4 @@ export default defineConfig({
     host: "0.0.0.0",
     port: 3000
   }
-})
+});
